Added all examples to regression testing

+++ ncpol2sdpa/sdp_relaxation.py
@@ -552,7 +552,7 @@ class SdpRelaxation(Relaxation):
         monomials = \
             pick_monomials_up_to_degree(all_monomials, max_localization_order)
         A = np.zeros(
-            (len(equalities) * len(monomials) * (len(monomials) + 1) / 2,
+            (int(len(equalities) * len(monomials) * (len(monomials) + 1) / 2),
              self.n_vars + 1))
         n_rows = 0
         for equality in equalities:
